Created new Decoy Sequence Validator
