+++ mzLib/Proteomics/Protein/Protein.cs
@@ -785,158 +785,6 @@ namespace Proteomics
 
         #endregion
 
-        /// <summary>
-        /// This function takes in a decoy protein and a list of forbidden sequences that the decoy
-        /// protein should not contain. Optionally, a list of the peptides within the base sequence
-        /// of the decoy protein that need to be scrambled can be passed as well. It will scramble the required sequences,
-        /// leaving cleavage sites intact. 
-        /// </summary>
-        /// <param name="originalDecoyProtein"> A Decoy protein to be cloned </param>
-        /// <param name="digestionParams"> Digestion parameters </param>
-        /// <param name="forbiddenSequences"> A HashSet of forbidden sequences that the decoy protein should not contain. Typically, a set of target base sequences </param>
-        /// <param name="sequencesToScramble"> Optional IEnumberable of sequences within the decoy protein that need to be replaced.
-        ///                                     If this is passed, only sequences within the IEnumerable will be replaced!!! </param>
-        /// <returns> A cloned copy of the decoy protein with a scrambled sequence </returns>
-        public static Protein ScrambleDecoyProteinSequence(
-            Protein originalDecoyProtein,
-            DigestionParams digestionParams,
-            HashSet<string> forbiddenSequences,
-            IEnumerable<string> sequencesToScramble = null)
-        {
-            // If no sequencesToScramble are passed in, we check to see if any 
-            // peptides in the decoy are forbidden sequences
-            sequencesToScramble = sequencesToScramble ?? originalDecoyProtein
-                .Digest(digestionParams, new List<Modification>(), new List<Modification>())
-                .Select(pep => pep.FullSequence)
-                .Where(forbiddenSequences.Contains);
-            if(sequencesToScramble.Count() == 0)
-            {
-                return originalDecoyProtein;
-            }
-
-            string scrambledProteinSequence = originalDecoyProtein.BaseSequence;
-            // Clone the original protein's modifications
-            var scrambledModificationDictionary = originalDecoyProtein.OriginalNonVariantModifications.ToDictionary(kvp => kvp.Key, kvp => kvp.Value);
-            Random rng = new Random(42);
-
-            // Start small and then go big. If we scramble a zero-missed cleavage peptide, but the missed cleavage peptide contains the previously scrambled peptide
-            // Then we can avoid unnecessary operations as the scrambledProteinSequence will no longer contain the longer sequence of the missed cleavage peptide
-            foreach(string peptideSequence in sequencesToScramble.OrderBy(seq => seq.Length))
-            {
-                if(scrambledProteinSequence.Contains(peptideSequence))
-                {
-                    string scrambledPeptideSequence = ScrambleSequence(peptideSequence, digestionParams.DigestionAgent.DigestionMotifs, rng,
-                        out var swappedArray);
-                    int scrambleAttempts = 1;
-
-                    // Try five times to scramble the peptide sequence without creating a forbidden sequence
-                    while(forbiddenSequences.Contains(scrambledPeptideSequence) & scrambleAttempts <= 5)
-                    {
-                        scrambledPeptideSequence = ScrambleSequence(peptideSequence, digestionParams.DigestionAgent.DigestionMotifs, rng,
-                            out swappedArray);
-                        scrambleAttempts++;
-                    }
-
-                    scrambledProteinSequence = scrambledProteinSequence.Replace(peptideSequence, scrambledPeptideSequence);
-
-                    if (!scrambledModificationDictionary.Any()) continue;
-
-                    // rearrange the modifications 
-                    foreach (int index in scrambledProteinSequence.IndexOfAll(scrambledPeptideSequence))
-                    {
-                        // Get mods that were affected by the scramble
-                        var relevantMods = scrambledModificationDictionary.Where(kvp => 
-                            kvp.Key >= index + 1 && kvp.Key < index + peptideSequence.Length + 1).ToList();
-
-                        // Modify the dictionary to reflect the new positions of the modifications
-                        foreach (var kvp in relevantMods)
-                        {
-                            int newKey = swappedArray[kvp.Key - 1 - index] + 1 + index;
-                            // To prevent collisions, we have to check if mods already exist at the new idx.
-                            if(scrambledModificationDictionary.TryGetValue(newKey, out var modsToSwap))
-                            {
-                                // If there are mods at the new idx, we swap the mods
-                                scrambledModificationDictionary[newKey] = kvp.Value;
-                                scrambledModificationDictionary[kvp.Key] = modsToSwap;
-                            }
-                            else
-                            {
-                                scrambledModificationDictionary.Add(newKey, kvp.Value);
-                                scrambledModificationDictionary.Remove(kvp.Key);
-                            }
-                        }
-                    }
-                }
-            }
-
-            Protein newProtein = originalDecoyProtein.CloneWithNewSequenceAndMods(scrambledProteinSequence, scrambledModificationDictionary) as Protein;
-
-            return newProtein;
-        }
-
-        /// <summary>
-        /// Scrambles a peptide sequence, preserving the position of any cleavage sites.
-        /// </summary>
-        /// <param name="swappedPositionArray">An array that maps the previous position (index) to the new position (value)</param>
-        public static string ScrambleSequence(string sequence, List<DigestionMotif> motifs, Random rng, out int[] swappedPositionArray)
-        {
-            // First, find the location of every cleavage motif. These sites shouldn't be scrambled.
-            HashSet<int> zeroBasedCleavageSitesLocations = new();
-            foreach (var motif in motifs)
-            {
-                for (int i = 0; i < sequence.Length; i++)
-                {
-                    (bool fits, bool prevents) = motif.Fits(sequence, i);
-                    if (fits && !prevents)
-                    {
-                        zeroBasedCleavageSitesLocations.Add(i);
-                    }
-                }
-            }
-
-            // Next, scramble the sequence using the Fisher-Yates shuffle algorithm.
-            char[] sequenceArray = sequence.ToCharArray();
-            // We're going to keep track of the positions of the characters in the original sequence,
-            // This will enable us to adjust the location of modifications that are present in the original sequence
-            // to the new scrambled sequence.
-            int[] tempPositionArray = Enumerable.Range(0, sequenceArray.Length).ToArray();
-            int n = sequenceArray.Length;
-            while(n > 1)
-            {
-                n--;
-                if(zeroBasedCleavageSitesLocations.Contains(n))
-                {
-                    // Leave the cleavage site in place
-                    continue;
-                }
-                int k = rng.Next(n + 1);
-                // don't swap the position of a cleavage site
-                while(zeroBasedCleavageSitesLocations.Contains(k))
-                {
-                    k = rng.Next(n + 1);
-                }
-
-                // rearrange the sequence array
-                char tempResidue = sequenceArray[k];
-                sequenceArray[k] = sequenceArray[n];
-                sequenceArray[n] = tempResidue;
-
-                // update the position array to represent the swaps
-                int tempPosition = tempPositionArray[k];
-                tempPositionArray[k] = tempPositionArray[n];
-                tempPositionArray[n] = tempPosition;
-            }
-
-            // This maps the previous position (index) to the new position (value)
-            swappedPositionArray = new int[tempPositionArray.Length];
-            for (int i = 0; i < tempPositionArray.Length; i++)
-            {
-                swappedPositionArray[tempPositionArray[i]] = i;
-            }
-
-            return new string(sequenceArray);
-        }
-
         public IBioPolymer CloneWithNewSequenceAndMods(string newBaseSequence, IDictionary<int, List<Modification>>? newMods = null)
         {
             // Create a new protein with the new base sequence and modifications
